requirements.txt
streamlit
playwright
psycopg2
SQLAlchemy
beautifulsoup4
python-dotenv
#playwright install